cleaned

+++ render.sh
@@ -1,14 +1,24 @@
 #!/usr/bin/env bash
 
+set -e
+
+#AUTHOR="Katabuchi, M."
+LASTNAME=Katabuchi
+
 pandoc ref.md \
   --citeproc \
   --csl=my_ref.csl \
 	--bibliography=ref.bib \
 	--pdf-engine=xelatex \
 	-s -o ref.tex; \
-	pandoc ref.tex -o ref_citeproc.md; \
-	rm -f ref_citeproc_edit.md
-	python ref.py "Katabuchi, M."
+	pandoc ref.tex -o ref_citeproc.md
+
+if [ -f ref_citeproc_edit.md ]; then
+  rm -f ref_citeproc_edit.md
+fi
+
+#python ref.py ${AUTHOR}
+python ref.py "Katabuchi, M."
 
 pandoc -s CV1.md ref_citeproc_edit.md CV2.md \
   --citeproc \
@@ -18,6 +28,6 @@ pandoc -s CV1.md ref_citeproc_edit.md CV2.md \
   --bibliography=ref.bib \
   --csl=my_ref.csl \
   -V papersize=letter \
-  -o Katabuchi_CV_raw.tex; \
-sed -e "s/enumerate/itemize/g" Katabuchi_CV_raw.tex > Katabuchi_CV.tex 
-context Katabuchi_CV.tex
+  -o ${LASTNAME}_CV_raw.tex; \
+  sed -e "s/enumerate/itemize/g" ${LASTNAME}_CV_raw.tex > ${LASTNAME}_CV.tex 
+context ${LASTNAME}_CV.tex
